+++ internal/httpclient/models/registration_request.go
@@ -13,7 +13,7 @@ import (
 	"github.com/go-openapi/validate"
 )
 
-// RegistrationRequest RegistrationRequest RegistrationRequest registration request
+// RegistrationRequest registration request
 // swagger:model registrationRequest
 type RegistrationRequest struct {
 
@@ -23,8 +23,6 @@ type RegistrationRequest struct {
 	// ExpiresAt is the time (UTC) when the request expires. If the user still wishes to log in,
 	// a new request has to be initiated.
 	// Format: date-time
-	// Format: date-time
-	// Format: date-time
 	ExpiresAt strfmt.DateTime `json:"expires_at,omitempty"`
 
 	// id
@@ -32,8 +30,6 @@ type RegistrationRequest struct {
 	ID UUID `json:"id,omitempty"`
 
 	// IssuedAt is the time (UTC) when the request occurred.
-	// Format: date-time
-	// Format: date-time
 	// Format: date-time
 	IssuedAt strfmt.DateTime `json:"issued_at,omitempty"`